refactor: STRING 객체의 상수 변수들 한 파일로 모음
