+++ src/BridgeMaker.js
@@ -1,15 +1,12 @@
-const STRING = {
-  DOWN_CHAR: 'D',
-  UP_CHAR: 'U',
-};
+const STRING = require('./Constant/Constant');
 
 const BridgeMaker = {
   makeBridge(size, generateRandomNumber) {
     const bridge = [];
     for (let i = 0; i < size; i += 1) {
       const randomNumber = generateRandomNumber();
-      if (randomNumber === 0) bridge.push(STRING.DOWN_CHAR);
-      if (randomNumber === 1) bridge.push(STRING.UP_CHAR);
+      if (randomNumber === 0) bridge.push(STRING.DOWN);
+      if (randomNumber === 1) bridge.push(STRING.UP);
     }
     return bridge;
   },
