+++ src/View/OutputView.js
@@ -2,24 +2,10 @@ const MissionUtils = require('@woowacourse/mission-utils');
 
 const { Console } = MissionUtils;
 
-const STRING = {
-  START_OUTPUT: '다리 건너기 게임을 시작합니다.',
-  UP_CHAR: 'U',
-  DOWN_CHAR: 'D',
-  O_CHAR: 'O',
-  X_CHAR: 'X',
-  EMPTY_STRING: '',
-  BLANK: ' ',
-  LEFT_WALL: '[',
-  RIGHT_WALL: ']',
-  MIDDLE_WALL: '|',
-  UP_INDEX: 0,
-  DOWN_INDEX: 1,
-  FINAL_GAME_RESULT: '\n최종 게임 결과',
-  GAME_SUCCESS: '\n게임 성공 여부: 성공',
-  GAME_FAILURE: '\n게임 성공 여부: 실패',
-  TOTAL_ATTEMPTS: '총 시도한 횟수: ',
-};
+const STRING = require('../Constant/Constant');
+
+const UP_INDEX = 0;
+const DOWN_INDEX = 1;
 
 const OutputView = {
   startOutput() {
@@ -27,17 +13,27 @@ const OutputView = {
   },
 
   printMap(bridge, location, moveSuccess) {
-    const map = [STRING.EMPTY_STRING, STRING.EMPTY_STRING];
-    map[STRING.UP_INDEX] += this.makeOneLineMap(bridge, location, moveSuccess, STRING.UP_CHAR);
-    map[STRING.DOWN_INDEX] += this.makeOneLineMap(bridge, location, moveSuccess, STRING.DOWN_CHAR);
-    Console.print(map[STRING.UP_INDEX]);
-    Console.print(map[STRING.DOWN_INDEX]);
+    const map = [STRING.EMPTY, STRING.EMPTY];
+    map[UP_INDEX] += this.makeOneLineMap(
+      bridge,
+      location,
+      moveSuccess,
+      STRING.UP
+    );
+    map[DOWN_INDEX] += this.makeOneLineMap(
+      bridge,
+      location,
+      moveSuccess,
+      STRING.DOWN
+    );
+    Console.print(map[UP_INDEX]);
+    Console.print(map[DOWN_INDEX]);
   },
 
   makeOneLineMap(bridge, location, moveSuccess, upOrDown) {
     let oneLineMap = STRING.LEFT_WALL + STRING.BLANK;
     for (let index = 0; index < location; index += 1) {
-      if (bridge[index] === upOrDown) oneLineMap += STRING.O_CHAR;
+      if (bridge[index] === upOrDown) oneLineMap += STRING.O;
       else oneLineMap += STRING.BLANK;
       oneLineMap += STRING.BLANK + STRING.MIDDLE_WALL + STRING.BLANK;
     }
@@ -48,11 +44,11 @@ const OutputView = {
 
   addLastMap(bridge, location, moveSuccess, upOrDown) {
     if (moveSuccess) {
-      if (bridge[location] === upOrDown) return STRING.O_CHAR;
+      if (bridge[location] === upOrDown) return STRING.O;
       return STRING.BLANK;
     }
     if (bridge[location] === upOrDown) return STRING.BLANK;
-    return STRING.X_CHAR;
+    return STRING.X;
   },
 
   printResult(bridge, location, isSuccess, attempts) {
